video user back-end done

+++ src/controllers/videoController.js
@@ -31,19 +31,28 @@ export const search = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found" });
   }
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
   return res.render("edit", { pageTitle: "Editing", video });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
+
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found" });
   }
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -83,7 +92,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video Not Found" });
+  }
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
   await Video.findByIdAndDelete(id);
-
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
